LayoutTests/fast/dom/HTMLImageElement/sizes/image-sizes-invalids-expected.txt

PASS document.getElementById(elementId).clientWidth is 800
PASS currentSrcFileName(elementId) is "image-set-2x.png"
PASS document.getElementById(elementId).clientWidth is 800
PASS currentSrcFileName(elementId) is "image-set-2x.png"
PASS document.getElementById(elementId).clientWidth is 800
PASS currentSrcFileName(elementId) is "image-set-2x.png"
PASS document.getElementById(elementId).clientWidth is 800
PASS currentSrcFileName(elementId) is "image-set-2x.png"
PASS document.getElementById(elementId).clientWidth is 800
PASS currentSrcFileName(elementId) is "image-set-2x.png"
PASS document.getElementById(elementId).clientWidth is 800
PASS currentSrcFileName(elementId) is "image-set-2x.png"
PASS document.getElementById(elementId).clientWidth is 800
PASS currentSrcFileName(elementId) is "image-set-2x.png"
PASS document.getElementById(elementId).clientWidth is 800
PASS currentSrcFileName(elementId) is "image-set-2x.png"
PASS document.getElementById(elementId).clientWidth is 800
PASS currentSrcFileName(elementId) is "image-set-2x.png"
PASS successfullyParsed is true

TEST COMPLETE
        
